# hw2/api/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
...
  /products:
    post:
      operationId: createProduct
      summary: "Создание товара (SELLER: свои, ADMIN: любые)"
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: "VALIDATION_ERROR"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "ACCESS_DENIED"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      operationId: listProducts
      summary: "Список товаров (все роли)"
      tags: [Products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{id}:
    get:
      operationId: getProduct
      summary: "Получение товара (все роли)"
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: "PRODUCT_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateProduct
      summary: "Обновление товара (SELLER: свои, ADMIN: любые)"
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: "VALIDATION_ERROR"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "ACCESS_DENIED"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "PRODUCT_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: "INVALID_STATE_TRANSITION"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteProduct
      summary: "Мягкое удаление (SELLER: свои, ADMIN: любые)"
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Товар архивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '403':
          description: "ACCESS_DENIED"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "PRODUCT_NOT_FOUND"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: "INVALID_STATE_TRANSITION"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /promo-codes:
    post:
      operationId: createPromoCode
      summary: "Создание промокода (SELLER, ADMIN)"
      tags: [PromoCodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          description: "VALIDATION_ERROR"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "ACCESS_DENIED"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Claims: user_id (UUID), role (USER/SELLER/ADMIN)"
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]
...
    DiscountType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]
...
    ProductCreate:
      type: object
      required: [name, price, stock, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
    ProductUpdate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductListResponse:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
...
    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: double
          minimum: 0.01
        min_order_amount:
          type: number
          format: double
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
    PromoCodeResponse:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: double
        min_order_amount:
          type: number
          format: double
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: |
            PRODUCT_NOT_FOUND, PRODUCT_INACTIVE, ORDER_NOT_FOUND,
            ORDER_LIMIT_EXCEEDED, ORDER_HAS_ACTIVE, INVALID_STATE_TRANSITION,
            INSUFFICIENT_STOCK, PROMO_CODE_INVALID, PROMO_CODE_MIN_AMOUNT,
            ORDER_OWNERSHIP_VIOLATION, VALIDATION_ERROR, ACCESS_DENIED,
            TOKEN_EXPIRED, TOKEN_INVALID, REFRESH_TOKEN_INVALID
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true